+++ pkg/listener/metrics/result_custom_test.go
@@ -12,12 +12,13 @@ import (
 	ioprometheusclient "github.com/prometheus/client_model/go"
 )
 
-func Test_SimpleClusterResultMetricGeneration(t *testing.T) {
-	gauge := metrics.RegisterSimpleClusterResultGauge("cluster_policy_report_simple_result")
+func Test_CustomResultMetricGeneration(t *testing.T) {
+	gauge := metrics.RegisterCustomResultGauge("policy_report_custom_result", []string{"namespace", "policy", "status", "source"})
 
 	report1 := report.PolicyReport{
 		ID:                "1",
 		Name:              "polr-test",
+		Namespace:         "test",
 		Summary:           report.Summary{Pass: 2, Fail: 1},
 		CreationTimestamp: time.Now(),
 		Results:           []report.Result{result1, result2, result3},
@@ -26,6 +27,7 @@ func Test_SimpleClusterResultMetricGeneration(t *testing.T) {
 	report2 := report.PolicyReport{
 		ID:                "1",
 		Name:              "polr-test",
+		Namespace:         "test",
 		Summary:           report.Summary{Pass: 0, Fail: 1},
 		CreationTimestamp: time.Now(),
 		Results:           []report.Result{result1, result3},
@@ -34,7 +36,7 @@ func Test_SimpleClusterResultMetricGeneration(t *testing.T) {
 	filter := metrics.NewResultFilter(validate.RuleSets{}, validate.RuleSets{}, validate.RuleSets{Exclude: []string{"disallow-policy"}}, validate.RuleSets{}, validate.RuleSets{})
 
 	t.Run("Added Metric", func(t *testing.T) {
-		handler := metrics.CreateSimpleClusterResultMetricsListener(filter, gauge)
+		handler := metrics.CreateCustomResultMetricsListener(filter, gauge, metrics.CreateLabelGenerator([]string{"namespace", "policy", "status", "source"}))
 		handler(report.LifecycleEvent{Type: report.Added, NewPolicyReport: report1, OldPolicyReport: report.PolicyReport{}})
 
 		metricFam, err := prometheus.DefaultGatherer.Gather()
@@ -42,16 +44,16 @@ func Test_SimpleClusterResultMetricGeneration(t *testing.T) {
 			t.Errorf("unexpected Error: %s", err)
 		}
 
-		results := findMetric(metricFam, "cluster_policy_report_simple_result")
+		results := findMetric(metricFam, "policy_report_custom_result")
 		if results == nil {
-			t.Fatalf("Metric not found: cluster_policy_report_simple_result")
+			t.Fatalf("Metric not found: policy_report_custom_result")
 		}
 
 		metrics := results.GetMetric()
-		if err = testSimpleClusterResultMetricLabels(metrics[0], result2, 1); err != nil {
+		if err = testCustomResultMetricLabels(metrics[0], result2, 1); err != nil {
 			t.Error(err)
 		}
-		if err = testSimpleClusterResultMetricLabels(metrics[1], result1, 1); err != nil {
+		if err = testCustomResultMetricLabels(metrics[1], result1, 1); err != nil {
 			t.Error(err)
 		}
 	})
@@ -59,7 +61,7 @@ func Test_SimpleClusterResultMetricGeneration(t *testing.T) {
 	t.Run("Modified Metric", func(t *testing.T) {
 		gauge.Reset()
 
-		handler := metrics.CreateSimpleClusterResultMetricsListener(filter, gauge)
+		handler := metrics.CreateCustomResultMetricsListener(filter, gauge, metrics.CreateLabelGenerator([]string{"namespace", "policy", "status", "source"}))
 		handler(report.LifecycleEvent{Type: report.Added, NewPolicyReport: report1, OldPolicyReport: report.PolicyReport{}})
 		handler(report.LifecycleEvent{Type: report.Updated, NewPolicyReport: report2, OldPolicyReport: report1})
 
@@ -68,16 +70,16 @@ func Test_SimpleClusterResultMetricGeneration(t *testing.T) {
 			t.Errorf("unexpected Error: %s", err)
 		}
 
-		results := findMetric(metricFam, "cluster_policy_report_simple_result")
+		results := findMetric(metricFam, "policy_report_custom_result")
 		if results == nil {
-			t.Fatalf("Metric not found: cluster_policy_report_simple_result")
+			t.Fatalf("Metric not found: policy_report_custom_result")
 		}
 
 		metrics := results.GetMetric()
-		if err = testSimpleClusterResultMetricLabels(metrics[0], result2, 0); err != nil {
+		if err = testCustomResultMetricLabels(metrics[0], result2, 0); err != nil {
 			t.Error(err)
 		}
-		if err = testSimpleClusterResultMetricLabels(metrics[1], result1, 1); err != nil {
+		if err = testCustomResultMetricLabels(metrics[1], result1, 1); err != nil {
 			t.Error(err)
 		}
 	})
@@ -85,7 +87,7 @@ func Test_SimpleClusterResultMetricGeneration(t *testing.T) {
 	t.Run("Deleted Metric", func(t *testing.T) {
 		gauge.Reset()
 
-		handler := metrics.CreateSimpleClusterResultMetricsListener(filter, gauge)
+		handler := metrics.CreateCustomResultMetricsListener(filter, gauge, metrics.CreateLabelGenerator([]string{"namespace", "policy", "status", "source"}))
 		handler(report.LifecycleEvent{Type: report.Added, NewPolicyReport: report1, OldPolicyReport: report.PolicyReport{}})
 		handler(report.LifecycleEvent{Type: report.Updated, NewPolicyReport: report2, OldPolicyReport: report1})
 		handler(report.LifecycleEvent{Type: report.Deleted, NewPolicyReport: report2, OldPolicyReport: report.PolicyReport{}})
@@ -95,57 +97,59 @@ func Test_SimpleClusterResultMetricGeneration(t *testing.T) {
 			t.Errorf("unexpected Error: %s", err)
 		}
 
-		results := findMetric(metricFam, "cluster_policy_report_simple_result")
+		results := findMetric(metricFam, "policy_report_custom_result")
 		if results == nil {
-			t.Fatalf("Metric not found: cluster_policy_report_simple_result")
+			t.Fatalf("Metric not found: policy_report_custom_result")
 		}
 
 		metrics := results.GetMetric()
-		if err = testSimpleClusterResultMetricLabels(metrics[0], result2, 0); err != nil {
+		if err = testCustomResultMetricLabels(metrics[0], result2, 0); err != nil {
 			t.Error(err)
 		}
-		if err = testSimpleClusterResultMetricLabels(metrics[1], result1, 0); err != nil {
+		if err = testCustomResultMetricLabels(metrics[1], result1, 0); err != nil {
 			t.Error(err)
 		}
 	})
 }
 
-func testSimpleClusterResultMetricLabels(metric *ioprometheusclient.Metric, result report.Result, expV float64) error {
-	if name := *metric.Label[0].Name; name != "category" {
+func testCustomResultMetricLabels(metric *ioprometheusclient.Metric, result report.Result, expVal float64) error {
+	var index int
+
+	if name := *metric.Label[index].Name; name != "namespace" {
 		return fmt.Errorf("unexpected Name Label: %s", name)
 	}
-	if value := *metric.Label[0].Value; value != result.Category {
-		return fmt.Errorf("unexpected Category Label Value: %s", value)
+	if value := *metric.Label[index].Value; value != result.Resource.Namespace {
+		return fmt.Errorf("unexpected Namespace Label Value: %s", value)
 	}
 
-	if name := *metric.Label[1].Name; name != "policy" {
+	index++
+
+	if name := *metric.Label[index].Name; name != "policy" {
 		return fmt.Errorf("unexpected Name Label: %s", name)
 	}
-	if value := *metric.Label[1].Value; value != result.Policy {
+	if value := *metric.Label[index].Value; value != result.Policy {
 		return fmt.Errorf("unexpected Policy Label Value: %s", value)
 	}
 
-	if name := *metric.Label[2].Name; name != "severity" {
-		return fmt.Errorf("unexpected Name Label: %s", name)
-	}
-	if value := *metric.Label[2].Value; value != result.Severity {
-		return fmt.Errorf("unexpected Severity Label Value: %s", value)
-	}
+	index++
 
-	if name := *metric.Label[3].Name; name != "source" {
+	if name := *metric.Label[index].Name; name != "source" {
 		return fmt.Errorf("unexpected Source Label: %s", name)
 	}
-	if value := *metric.Label[3].Value; value != result.Source {
+	if value := *metric.Label[index].Value; value != result.Source {
 		return fmt.Errorf("unexpected Source Label Value: %s", value)
 	}
 
-	if name := *metric.Label[4].Name; name != "status" {
+	index++
+
+	if name := *metric.Label[index].Name; name != "status" {
 		return fmt.Errorf("unexpected Name Label: %s", name)
 	}
-	if value := *metric.Label[4].Value; value != result.Status {
+	if value := *metric.Label[index].Value; value != result.Status {
 		return fmt.Errorf("unexpected Status Label Value: %s", value)
 	}
-	if value := metric.Gauge.GetValue(); value != expV {
+
+	if value := metric.Gauge.GetValue(); value != expVal {
 		return fmt.Errorf("unexpected Metric Value: %v", value)
 	}